+++ examples/android/android-test/src/cucumber/example/android/test/CucumberActivity.java
@@ -2,7 +2,6 @@ package cucumber.example.android.test;
 
 import android.app.Activity;
 import android.os.Bundle;
-import cucumber.android.test.R;
 
 public class CucumberActivity extends Activity {
     public void onCreate(Bundle savedInstanceState) {