+++ examples/android/android-test/src/cucumber/example/android/test/CucumberActivitySteps.java
@@ -3,6 +3,7 @@ package cucumber.example.android.test;
 import android.app.Instrumentation;
 import android.test.ActivityInstrumentationTestCase2;
 import android.util.Log;
+import cucumber.api.CucumberOptions;
 import cucumber.api.android.CucumberInstrumentation;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
@@ -10,8 +11,24 @@ import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 
-// Glue code classes need to extend Android test classes
-// in order to have access to Context and Instrumentation.
+/**
+ * We extend ActivityInstrumentationTestCase2 in order to have access to methods like getActivity
+ * and getInstrumentation. Depending on what methods we are going to need, we can put our
+ * step definitions inside classes extending any of the following Android test classes:
+ * <p/>
+ * ActivityInstrumentationTestCase2
+ * InstrumentationTestCase
+ * AndroidTestCase
+ * <p/>
+ * The CucumberOptions annotation is mandatory for exactly one of the classes in the test project.
+ * Only the first annotated class that is found will be used, others are ignored. If no class is
+ * annotated, an exception is thrown.
+ * <p/>
+ * The options need to at least specify features = "features". The default value that is set by
+ * Cucumber internally does not work because features are not on the classpath under Android.
+ * Features must be placed inside assets/features/ of the test project (or a subdirectory thereof).
+ */
+@CucumberOptions(features = "features")
 public class CucumberActivitySteps extends ActivityInstrumentationTestCase2<CucumberActivity> {
     private int steps;