+++ examples/android/cukeulator-test/src/cucumber/example/android/cukeulator/test/CalculatorActivitySteps.java
@@ -2,6 +2,7 @@ package cucumber.example.android.cukeulator.test;
 
 import android.test.ActivityInstrumentationTestCase2;
 import android.widget.TextView;
+import cucumber.api.CucumberOptions;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
 import cucumber.api.java.en.Given;
@@ -12,29 +13,38 @@ import cucumber.example.android.cukeulator.R;
 
 import static cucumber.example.android.cukeulator.test.Utils.clickOnView;
 
+/**
+ * We extend ActivityInstrumentationTestCase2 in order to have access to methods like getActivity
+ * and getInstrumentation. Depending on what methods we are going to need, we can put our
+ * step definitions inside classes extending any of the following Android test classes:
+ * <p/>
+ * ActivityInstrumentationTestCase2
+ * InstrumentationTestCase
+ * AndroidTestCase
+ * <p/>
+ * The CucumberOptions annotation is mandatory for exactly one of the classes in the test project.
+ * Only the first annotated class that is found will be used, others are ignored. If no class is
+ * annotated, an exception is thrown.
+ * <p/>
+ * The options need to at least specify features = "features". Features must be placed inside
+ * assets/features/ of the test project (or a subdirectory thereof).
+ */
+@CucumberOptions(features = "features")
 public class CalculatorActivitySteps extends ActivityInstrumentationTestCase2<CalculatorActivity> {
-    private CalculatorActivity activity;
 
     public CalculatorActivitySteps() {
         super(CalculatorActivity.class);
     }
 
-    @Before
-    public void before() {
-    }
-
-    @After
-    public void after() {
-    }
-
     @Given("^I have a CalculatorActivity$")
     public void I_have_a_CalculatorActivity() {
-        activity = getActivity();
-        assertNotNull(activity);
+        assertNotNull(getActivity());
     }
 
     @When("^I press (\\d)$")
     public void I_press_d(int d) {
+        CalculatorActivity activity = getActivity();
+
         switch (d) {
             case 0:
                 clickOnView(activity, R.id.btn_d_0);
@@ -71,6 +81,8 @@ public class CalculatorActivitySteps extends ActivityInstrumentationTestCase2<Ca
 
     @When("^I press ([+–x\\/=])$")
     public void I_press_op(char op) {
+        CalculatorActivity activity = getActivity();
+
         switch (op) {
             case '+':
                 clickOnView(activity, R.id.btn_op_add);
@@ -92,7 +104,7 @@ public class CalculatorActivitySteps extends ActivityInstrumentationTestCase2<Ca
 
     @Then("^I should see (\\S+) on the display$")
     public void I_should_see_s_on_the_display(String s) {
-        TextView display = (TextView) activity.findViewById(R.id.txt_calc_display);
+        TextView display = (TextView) getActivity().findViewById(R.id.txt_calc_display);
         String displayed_result = display.getText().toString();
         assertEquals(s, displayed_result);
     }